pyproject.toml
[project]
name = "black-bull-stats"
version = "0.1.0"
description = "Simple app to display Black Bulls FC team stats"
authors = [
    {name = "stackswithans",email = "stexor12@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "streamlit (>=1.41.1,<2.0.0)",
    "st-gsheets-connection (>=0.1.0,<0.2.0)",
    "plotly (>=5.24.1,<6.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
